+++ ChatModuleDemo/TCPClient/clientMain.cpp
@@ -12,20 +12,20 @@ int main(int arg, char** argv)
 	int result = 0;
 	const int bufSize = 512;
 
-	//Other class instances
+	// Other class instances
 	cNetworkManager networkManager;
 	cUserInterface userInterface;
 	sChatMessage message;
 
 	cBuffer buffer(bufSize);
 
-	// Initialize the Windows Sockets API (WSA).
+	// Initialize the Windows Sockets API (WSA)
 	result = networkManager.InitializeWSA();
 
 	if (result != 0)
 		return 1;
 
-	// Initialize address information.
+	// Initialize address information
 	result = networkManager.InitializeAddrInfo("127.0.0.1", DEFAULT_PORT);
 
 	if (result != 0)
@@ -33,13 +33,13 @@ int main(int arg, char** argv)
 
 	SOCKET serverSocket;
 
-	// Create a socket for the server connection.
+	// Create a socket for the server connection
 	result = networkManager.CreateSocket(serverSocket);
 
 	if (result != 0)
 		return 1;
 
-	// Connect to the server.
+	// Connect to the server
 	result = networkManager.ConnectSocket(serverSocket);
 
 	if (result != 0)
@@ -48,7 +48,7 @@ int main(int arg, char** argv)
 	//FD_SET for socket operations				
 	FD_SET socketsReadyForReading;
 
-	// timeval to prevent select from waiting forever.
+	// timeval to prevent select from waiting forever
 	timeval tv;
 	tv.tv_sec = 0;
 	tv.tv_usec = 0;
@@ -87,9 +87,9 @@ int main(int arg, char** argv)
 			continue;
 		else if (result == 0)
 		{
-			// Process incoming messages from server.
+			// Process incoming messages from server
 			uint32_t packetSize = buffer.ReadUInt32BE();
-			uint32_t messageType = buffer.ReadUInt32BE();
+			uint32_t messageType = buffer.ReadUInt16BE();
 
 			if (buffer.mBufferData.size() >= packetSize)
 			{
@@ -113,7 +113,7 @@ int main(int arg, char** argv)
 		buffer.mBufferData.resize(512);
 	}
 
-	// Clean up and close the connection.
+	// Clean up and close the connection
 	networkManager.CleanSocket(serverSocket, networkManager.info);
 
 	return 0;
